order_product_api/serializers.py
from rest_framework import serializers
from order_product.models import OrderProduct


class OrderProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = OrderProduct
        fields = '__all__'

